requirements.txt
PyQt6>=6.4.0
openpyxl>=3.1.0
pytesseract>=0.3.10
Pillow>=10.0.0
pyinstaller>=6.0.0
